Day 9/Medical-Career-Mentor-Final/requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
pydot
graphviz